requirements.txt
python-telegram-bot[job-queue]==20.8
python-dotenv==1.0.1
selenium>=4.0.0
webdriver-manager>=4.0.0 